Concept of Bar Graphs
Here I have taken a very popular Titanic datasets and I have also done analysis on that datasets and also shown the bar graph concept on a single variable i.e. how we could take the unique values in the X-axis and and we could then plotted there value counts on the Y-axis and on the basis of that we could easily differentiate between two unique categorical variables on the basis of stats as bar graphs are also used for numerical variables but it is more suitable for categorical variable.