Concept of Histograms
Now here I have shown how we could explore the data using exploratory data analysis and then how we could plot the graphs for a single variable using Histogram i.e. Histogram is generally used to show how much amount or percent of data i.e. frequency is there for a particular unique value in a particular column i.e. it is generally used to tell us about the frequency of the unique values in a variable.