Sub-setting or Boolean-Indexing
Here I have shown how to do selection of data from a dataset on the basis of variables present in the datasets i.e. selection of data on the basis of [] operator,loc function and on the basis of condition with one argument and two argument and also used this sub-setting in the group-by part.