Concept of Data Cleaning
Here I have shown how we could do cleaning of a variable first then we could do the analysis on the required data i.e. if suppose a particular variable on which we have to do the analysis but that particular variable is not in a required form to do analysis so first we generally used to clean that particular variable and then do the analysis so data cleaning is somehow cleaning a particular variable on the basis of string functions and some how change is done on that particular variable values but the change does not effect the datasets.