Checking of Correlation
So after completing the Analysis part on the datasets I have came across to show the concept of correlation on the datasets i.e. here I will tell how I have check about one of the most important term regarding i.e.correlation which means to check whether dependent variable is related or highly correlated to which independent variable but before checking the correlation we have to do data pre-processing which contains finding and replacing of missing values as well as converting the categorical data into continuous form using various libraries after that once the data preparation is done then finally check the assumptions of the Model which we are going to fit in the datasets Since here I have used the concept of linear regression so I have checked two very important assumptions Multic-colinearity and Linearity and then finally apply the model and try to check the approximation i.e. R-square between the actual dependent and predicated dependent variable and also check the residuals or we can say Outliers which is also known as the error rate present in the predicated part.