Linearity and Final Model Prediction
At last I have shown linearity concept where I have shown to which independent variable the dependent variable depends the most and then we will finally apply the linear Regression as the model on the datasets and then we will see some important stats like R-square which will check the approximation between the actual and predicated dependent variable and we will also find out the residuals and outliers which act as an error rate in the model prediction and we will also see the accuracy in the model prediction and finally we will plot the graph of the linear Regression through which we will try to depict or analyze about the analysis part i.e. We will try to distinguish between the Good and Bad users according to the various anomalies.